+++ src/icloudpd/config.py
@@ -50,6 +50,7 @@ class _DefaultConfig:
     skip_created_before: datetime.datetime | datetime.timedelta | None
     skip_created_after: datetime.datetime | datetime.timedelta | None
     skip_photos: bool
+    retrieve_all_first: bool
 
 
 @dataclass(kw_only=True)